+++ src/main/java/com/tarea3adtraullg/proyecto_pokemon/repositorios/RepoCarnet.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.tarea3adtraullg.proyecto_pokemon.entidades.Carnet;
 
-public interface RepoCarnet extends JpaRepository<Carnet, Integer> {
+public interface RepoCarnet extends JpaRepository<Carnet, Long> {
 
 }
